Move WebcamSaveBtn out of controls.tsx
