+++ src/controls/controls.tsx
@@ -12,9 +12,7 @@ import {
 import { BotState } from "../devices/interfaces";
 import { connect } from "react-redux";
 import { Everything } from "../interfaces";
-import { DeviceAccountSettings } from "../devices/interfaces";
-import { put } from "axios";
-
+import { WebcamSaveBtn } from "./webcam_save_btn";
 
 interface AxisInputBoxProps {
   bot: BotState;
@@ -335,46 +333,6 @@ export class Controls extends React.Component<Everything, any> {
     );
   }
 };
-
-interface Blah {
-  dispatch: Function;
-  apiUrl: string;
-  webcam_url: string;
-};
-
-let saveWebcamUrl = ({dispatch, apiUrl, webcam_url}: Blah) => () => {
-  return dispatch(function() {
-    put<DeviceAccountSettings>(apiUrl + "/api/device", { webcam_url })
-      .then(res => dispatch({ type: "REPLACE_DEVICE_ACCOUNT_INFO", payload: res.data }))
-      .catch((payload) => dispatch({ type: "DEVICE_ACCOUNT_ERR", payload }));
-  });
-};
-
-interface WebcamSaveBtnProps {
-  dispatch: Function;
-  webcamUrl: string;
-  apiUrl: string;
-  dirty: boolean;
-}
-
-function WebcamSaveBtn({dispatch, webcamUrl, apiUrl, dirty}) {
-  if (dirty) {
-    return <button type="button"
-                   className="green button-like widget-control"
-                   onClick={
-                     saveWebcamUrl({
-                       dispatch,
-                       apiUrl,
-                       webcam_url: webcamUrl
-                      })
-                   }>
-      SAVE
-    </button>;
-
-  } else {
-    return <span></span>;
-  };
-}
 
 const showUrl = (url: string, dirty: boolean) => {
   if (dirty) {
